Add support for Zephyr's native test environment
Zephyr's native test environment is a setup
which allows building the embedded SW with the Zephyr OS
and HW models as a Linux executable.

When building for this target the Zephyr integration
will set METAL_PROCESSOR_HEADERS_ON_SYSTEM.

We need to provide appropriate headers
for this environment.
The "metal_cpu_yield()" call, which in other
architectures is handled as either a hint for the CPU to yield
to another thread or do nothing, and therefore
busy waiting, needs replacing with a call out to
the Zephyr API which will cause the CPU to waste 1 microsecond,
therefore enabling the other CPU in the system to ready
whatever we are waiting for in the meanwhile.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>

(f) Add support for Zephyr's POSIX architecture

(f) Add support for Zephyr's native